Fix chart rendering in performance-management
The fillBars() function was refactored a couple times for D3 to
properly select the barchart in the element's template. One of those
refactorings was not completely removed, resulting in a reference to
an undefined variable in fillBars(), causing the drill-down in the
bar chart to fail and show up as blank.